feat: using constants for skills
