+++ app/data/constants/skills.ts
@@ -1,0 +1,52 @@
+import CSSIcon from "@/app/components/icons/skills/CSSIcon";
+import Git from "@/app/components/icons/skills/Git";
+import Java from "@/app/components/icons/skills/Java";
+import JavaScript from "@/app/components/icons/skills/JavaScript";
+import NextJS from "@/app/components/icons/skills/NextJS";
+import NodeJS from "@/app/components/icons/skills/NodeJS";
+import ReactIcon from "@/app/components/icons/skills/ReactIcon";
+import TailwindCSS from "@/app/components/icons/skills/TailwindCSS";
+import TypeScript from "@/app/components/icons/skills/TypeScript";
+import VSCode from "@/app/components/icons/skills/VSCode";
+
+const skillItems = [
+  { name: "JavaScript", Icon: JavaScript },
+  {
+    name: "TypeScript",
+    Icon: TypeScript,
+  },
+  {
+    name: "React",
+    Icon: ReactIcon,
+  },
+  {
+    name: "Next.js",
+    Icon: NextJS,
+  },
+  {
+    name: "CSS",
+    Icon: CSSIcon,
+  },
+  {
+    name: "Tailwind",
+    Icon: TailwindCSS,
+  },
+  {
+    name: "Node.js",
+    Icon: NodeJS,
+  },
+  {
+    name: "Java",
+    Icon: Java,
+  },
+  {
+    name: "Git",
+    Icon: Git,
+  },
+  {
+    name: "VS Code",
+    Icon: VSCode,
+  },
+];
+
+export default skillItems;
